# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ExamPortal API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface RegisterRequest {
  username: string;
  password: string;
  confirmPassword: string;
}

export interface ChangePasswordRequest {
  currentPassword: string;
  newPassword: string;
}

export interface UserInfo {
  username: string;
  isAdmin: boolean;
  joinDate: string;
  mustChangePassword: boolean;
}

export interface AuthResponse {
  user: UserInfo;
  mustChangePassword: boolean;
}

export interface Exam {
  id: number;
  title: string;
  department: string;
  countdown: number;
  fileName?: string;
  content?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateExamRequest {
  title: string;
  department: string;
  countdown: number;
  fileName?: string;
  content?: string;
}

export interface UpdateExamRequest {
  title?: string;
  department?: string;
  countdown?: number;
  fileName?: string;
  content?: string;
}

export interface Department {
  id: number;
  name: string;
}

export interface CreateDepartmentRequest {
  name: string;
}

export interface Message {
  id: number;
  sender: string;
  receiver: string;
  text: string;
  time: string;
  read: boolean;
}

export interface SendMessageRequest {
  receiver: string;
  text: string;
}

export interface UnreadCountResponse {
  count: number;
}

export interface MarkReadRequest {
  sender: string;
}

export type GetMessagesParams = {
  otherUser?: string;
};