# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ExamPortal API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: exams
    description: Exam management
  - name: departments
    description: Department management
  - name: messages
    description: Messaging
  - name: users
    description: User management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /exams:
    get:
      operationId: getExams
      tags: [exams]
      summary: Get all exams
      responses:
        "200":
          description: List of exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
    post:
      operationId: createExam
      tags: [exams]
      summary: Create exam (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExamRequest"
      responses:
        "201":
          description: Exam created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"

  /exams/{id}:
    get:
      operationId: getExam
      tags: [exams]
      summary: Get exam by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
    put:
      operationId: updateExam
      tags: [exams]
      summary: Update exam (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExamRequest"
      responses:
        "200":
          description: Updated exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
    delete:
      operationId: deleteExam
      tags: [exams]
      summary: Delete exam (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /departments:
    get:
      operationId: getDepartments
      tags: [departments]
      summary: Get all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create department (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"

  /departments/{id}:
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete department (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages for current user
      parameters:
        - name: otherUser
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /messages/unread-count:
    get:
      operationId: getUnreadCount
      tags: [messages]
      summary: Get unread message count
      responses:
        "200":
          description: Count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnreadCountResponse"

  /messages/mark-read:
    post:
      operationId: markMessagesRead
      tags: [messages]
      summary: Mark messages as read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkReadRequest"
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Get all users (admin)
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserInfo"

  /users/{username}:
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin)
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /events:
    get:
      operationId: getEvents
      tags: [health]
      summary: Server-Sent Events stream for real-time updates
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
      required:
        - username
        - password
        - confirmPassword

    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserInfo"
        mustChangePassword:
          type: boolean
      required:
        - user
        - mustChangePassword

    UserInfo:
      type: object
      properties:
        username:
          type: string
        isAdmin:
          type: boolean
        joinDate:
          type: string
        mustChangePassword:
          type: boolean
      required:
        - username
        - isAdmin
        - joinDate
        - mustChangePassword

    Exam:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        department:
          type: string
        countdown:
          type: integer
        fileName:
          type: string
        content:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - department
        - countdown
        - createdAt
        - updatedAt

    CreateExamRequest:
      type: object
      properties:
        title:
          type: string
        department:
          type: string
        countdown:
          type: integer
        fileName:
          type: string
        content:
          type: string
      required:
        - title
        - department
        - countdown

    UpdateExamRequest:
      type: object
      properties:
        title:
          type: string
        department:
          type: string
        countdown:
          type: integer
        fileName:
          type: string
        content:
          type: string

    Department:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
        - id
        - name

    CreateDepartmentRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name

    Message:
      type: object
      properties:
        id:
          type: integer
        sender:
          type: string
        receiver:
          type: string
        text:
          type: string
        time:
          type: string
        read:
          type: boolean
      required:
        - id
        - sender
        - receiver
        - text
        - time
        - read

    SendMessageRequest:
      type: object
      properties:
        receiver:
          type: string
        text:
          type: string
      required:
        - receiver
        - text

    UnreadCountResponse:
      type: object
      properties:
        count:
          type: integer
      required:
        - count

    MarkReadRequest:
      type: object
      properties:
        sender:
          type: string
      required:
        - sender