# lib/api-zod/src/generated/types/createExamRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ExamPortal API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateExamRequest {
  title: string;
  department: string;
  countdown: number;
  fileName?: string;
  content?: string;
}