# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ExamPortal API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  sender: string;
  receiver: string;
  text: string;
  time: string;
  read: boolean;
}