# lib/api-zod/src/generated/types/userInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ExamPortal API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserInfo {
  username: string;
  isAdmin: boolean;
  joinDate: string;
  mustChangePassword: boolean;
}